feat(*): Refactor to use OpenCTI templates
