+++ .github/workflows/release.yml
@@ -84,7 +84,7 @@ jobs:
 
       - name: Create zip for OpenCTI pull request
         run: |
-          zip -r ${{ env.VERSION_SOURCES }}.zip ${{ env.VERSION_SOURCES }}/ -x '${{ env.VERSION_SOURCES }}/.git*'    
+          zip -r ${{ env.VERSION_SOURCES }}.zip ${{ env.VERSION_SOURCES }}/ -x '${{ env.VERSION_SOURCES }}/.git*' -x '${{ env.VERSION_SOURCES }}/docs' -x '${{ env.VERSION_SOURCES }}/dev' 
 
       - name: Create Tag ${{ github.event.inputs.tag_name }}
         uses: actions/github-script@v7
