+++ .github/workflows/unittests.yml
@@ -33,7 +33,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install -r src/test-requirements.txt
+        python -m pip install -r tests/test-requirements.txt
 
     - name: Tests
       run: |
