+++ .gitignore
@@ -1,3 +1,3 @@
-config.yml
+src/config.yml
 __pycache__
 .python-version
