+++ src/crowdsec/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """CrowdSec external import connector module."""
 
-from .importer import CrowdSecImporter
+from .connector import CrowdSecImporter
 
 __all__ = ["CrowdSecImporter"]
