--- src/test-requirements.txt
@@ -1,3 +1,0 @@
--r ./requirements.txt
-pytest
-pytest-mock
