+++ tests/resources/stix_observable.json
